api-tgr/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
pymysql
mysql-connector-python
patentes-vehiculares-chile
python-jose[cryptography]
passlib[bcrypt]
python-multipart
rut-chile